aarch64: Remove unnecessary saves on calls with differing ABIs (#6303)
* aarch64: Remove unnecessary saves on calls with differing ABIs

The AArch64 AAPCS calling convention, which all our backends use for
register allocation, indicates that the low 64-bits of the v8-v15
registers are all caller-saved. Cranelift doesn't track precisely which
registers are used, however, so it says the entire register is a
caller-saved register. This currently interacts poorly where a call from
one function to another where the ABIs differ forces the caller to save
all of v8-v15 in the prologue and restore it in the epilogue.

Currently in all cases, however, this isn't actually necessary. The
AArch64 backend also has an optimization where if both the caller and
the callee are using the same ABI then the clobbers of a `call`
instruction are not counted in the clobber set for the function since
nothing new can be clobbered. This way if `v8` is never used, for
example, it's not considered clobbered and it's not saved. This logic,
however, is comparing ABIs exactly which means that different names for
the same ABI, which don't differ in register allocation, force saves to
happen.

This comes up with trampolines generated by Wasmtime where the calling
convention of the trampoline is `WasmtimeSystemV`, for example, where
the callee (a wasm function) is `Fast`. Because these differ it means
that all trampolines are generating saves/restores of registers, despite
the actual underlying calling convention being the same.

This commit updates the optimization that skips including a `call`
instruction in the clobber set by comparing the caller and callee's ABI
clobber sets. If both clobber the same registers then for the purposes
of clobbers it's as-if they were the same ABI, so the `call` can be
skipped.

Overall this removes unnecessary saves/restores in trampolines generated
by Cranelift and shrinks the size of spidermonkey.wasm by 2% after #6262.

* Use a subset check instead