aarch64: Fix usage of `load_addr` in `inst.isle` (#6314)
This fixes an issue in the AArch64 backend where a `load_addr` helper
was used exclusively for lowering `splat`-of-a-loaded-address. This
helper expanded in some cases to a pseudo-`LoadAddr` instruction but the
lowering of this instruction doesn't actually exhaustively handle all
`AMode` values.

The fix in this commit is to remove the `load_addr` helper altogether to
remove the need to go from an `AMode` back to a `Reg`, instead going
directly from an address to a register. The one small wrinkle is a small
helper now to add the immediate offset to the address register, but
that's not too too bad to write.

By avoiding the `LoadAddr` instruction the unimplemented cases aren't
hit, so the codegen issue should be fixed.