test: add cases from `wasi-testsuite` (#6341)
* wasi: add the `wasi-testsuite` tests for wasi-common

As described [here], this uses the `prod/testsuite-base` branch in which
the tests are built as `.wasm` files.

[here]: https://github.com/WebAssembly/wasi-testsuite/#getting-started

* chore: update `walkdir` everywhere to its latest version

This is done in order to use it for `wasi_testsuite` testing.

* vet: extend `walkdir`'s exemption

* test: factor out `get_wasmtime_command`

This will be helpful for `wasi_testsuite` testing.

* test: use all `wasi-testsuite` test cases

This change alters the `wasi_testsuite` test to run all of the available
test cases in [wasi-testsuite]. This involved making the test runner a
bit more robust to the various shapes of JSON specifications in that
project. Unfortunately, the `wasi_testsuite` test fails some of the
cases, so I added a `WASI_COMMON_IGNORE_LIST` to avoid these
temporarily. (This may remind some of the Wasm testsuite ignore lists in
Cranelift; those relied on `build.rs` to create a `#[test]` for each
test case, which I felt is not yet needed here).

It's unclear to me why the tests are failing. It could be because:
- wasi-common has a bug
- wasi-testsuite overspecifies (or incorrectly specifies) a test
- the test runner incorrectly configures Wasmtime's CLI execution.

But this change makes it easier to resolve this. Remove the file from
`WASI_COMMON_IGNORE_LIST` and run `cargo test wasi_testsuite --
--nocapture`. The printed output will show the expected result, the
actual result, and a command to replicate the failure from the command
line.

[wasi-testsuite]: https://github.com/WebAssembly/wasi-testsuite

* review: add "shrinking" comment