Remove the initializer from a global's type information (#6349)
* Remove the initializer from a global's type information

This commit removes the `Global::initializer` field to instead only
store type information about the global rather than its initialization
state. Instead now the initializer is stored separately from the type of
the global, and only for defined globals. This removes the need in a few
locations to synthesize dummy initializers.

* Actually delete what I intended to delete

* Simplify global initializer loop