Cranelift/ABI: catch overflow in layout code with large stack slots. (#6434)
* Filetest: allow "expect-fail" compile tests.

This mode allows us to test that we properly error out of compilation
situations where compilation should not be possible. It does not assert
anything about the specific error produced, just that a compile error
occurred.

* Add filetest for #6431.

* Cranelift/ABI: catch overflow in layout code with large stack slots.

This PR catches and successfully errors out of compilation a situation
where stackslots exceed the `u32` (4GB) range of the stack-frame layout
computation.

Note that this is not reachable from Wasmtime (it does not use
stackslots) so does not have security implications there, but could
cause issues in other compilers (e.g. cg\_clif) that pass arbitrary
stackslot sizes to Cranelift.