x64: Add memory operand support to EVEX instructions (#6416)
* x64: Add memory operand support to EVEX instructions

Currently load-sinking is enabled for EVEX instructions (aka AVX512
instructions) but the encoding of these instructions is a `todo!()`
which can cause a panic for some wasms if the right features are
enabled. This commit fills out the support for memory operands in the
same manner as VEX-encoded instructions. The main stickler here was that
EVEX instructions always use a scaled 8-bit offset which needed extra
handling to ensure that the correct offset is emitted.

* Pass scaling to `Imm` construction

* Review comments

* Fix tests

* Fix evex benchmark